Changed %zu to %lld in test-* (%zu doesn't work in vc)

+++ test/test-tightdb/test-tightdb.cpp
@@ -130,7 +130,7 @@ int main() {
 	}
 	table.Add(0, "abcde", 100, Wed);
 
-	printf("Memory usage: %zu bytes\n", GetMemUsage());
+	printf("Memory usage: %lld bytes\n", (int64_t)GetMemUsage()); // %zu doesn't work in vc
 
 	UnitTest::Timer timer;
 
@@ -192,7 +192,7 @@ int main() {
 		printf("Add index: %dms\n", search_time);
 	}
 
-	printf("Memory usage2: %zu bytes\n", GetMemUsage());
+	printf("Memory usage2: %lld bytes\n", (int64_t)GetMemUsage()); // %zu doesn't work in vc
 
 	// Search with index
 	{
